factory-ai-vision/EdgeSolution/modules/TrainingModule/backend/requirements/base.txt
django==3.0.8
djangorestframework==3.11.0
drf-yasg==1.17.1
drf-url-filters==0.5.1
numpy
pillow
pyzmq
azure-cognitiveservices-vision-customvision==1.0.0
requests
django-cors-headers
azure-iot-device~=2.0.0
python-dateutil
opencensus==0.7.7
opencensus-context==0.1.1
opencensus-ext-azure==1.0.2
opencensus-ext-django==0.7.2
opencensus-ext-stackdriver==0.7.2
channels==2.4.0
django-cleanup
